Create find.py
find.py assignment - just need to finish the recursive part! It needs to generate a list of available folders on a path, then do a loop over that list where each iteration performs a recursive call on the folder path. The file globbing is already implemented, so we don't need to check for plain files.